Renamed some classes and added cryptation
