+++ src/main/webapp/signIn.jsp
@@ -4,7 +4,7 @@
 
 <sy:layout title="Symposium-Login">
     <jsp:body>
-    <form action="GuestController?action=signin" method="post" data-abide novalidate>
+    <form action="GuestController?action=signIn" method="post" data-abide novalidate>
       <div data-abide-error class="alert callout" style="display: none;">
     <p><i class="fi-alert"></i> Ricontrolla la form</p></div>
     <c:if test="${wrongCredentials!=null}">
